Example volleyball match scout files

Example volleyball match scout files

Usage

ovdata_example(choice = "190301_kats_beds", as = "path", simplify = TRUE)

Arguments

choice

string: one or more files to return

"190301_kats_beds" - the full DataVolley file from a match between GKS Katowice and MKS Bedzin during the 2018/19 Polish Plus Liga (file courtesy Mark Lebedew)

"190301_kats_beds-clip" - a single rally from the "190301_kats_beds" DataVolley file. This rally corresponds to the video clip contained in ovdata_example_video("190301_kats_beds")

"PM06" - (DataVolley) the December 2012 men's Slovenian national championship semifinal between ACH Volley and Maribor (file from http://www.odbojka.si/)

"mlafin_braslovce_nkbm" - (DataVolley) the January 2015 Slovenian junior women's final between Braslovče and Nova KBM Branik (file from http://www.odbojka.si/)

"clickscout" - an example Click & Scout DataVolley file

"2017_AVL_mens_HEAT_vs_UTSSU" - (Perana/VBStats) Men's Australian Volleyball League 2017: Canberra Heat vs UTSSU (file courtesy Chau Le)

"2017_AVL_womens_HEAT_vs_UTSSU" - (Perana/VBStats) Women's Australian Volleyball League 2017: Canberra Heat vs UTSSU (file courtesy Chau Le)

"stuttgart_schwerin_2018" - (DataVolley) the 2018 women's final (first of 3) from the German Bundesliga, Allianz MTV Stuttgart vs SSC Palmberg Schwerin (file courtesy Michael Mattes)

"DCup7" - the 2020 Austrian Women's Volley Cup played between Hartberg and UVC Graz (file from https://www.volleynet.at/dvdownload/information/f-Damen/)

"NCA-CUB" - Nicaragua vs Cuba women from the Pan American Cup, August 2022 (vsm format, courtesy Christophe Elek)

as

string: either "path" (return the path to the file) or "parsed" (parse the file into an R data structure using datavolley::dv_read() or peranavolley::pv_read()

simplify

logical: by default if as is "parsed", the returned list is the same length as choice where each entry is a parsed object. If simplify is TRUE and we have asked for a single file (i.e. length(choice) == 1, then return just that object (not a list containing that object)

Value

If as is "path", a character vector with the file path(s), otherwise a list

See also

Examples

myfile <- ovdata_example()
x <- dv_read(myfile)
summary(x)
#> Match summary:
#> Date: 2019-01-03
#> League: Plus Liga 2018/2019 - Plus Liga 2018/20189- Faza Zasadnicza
#> Teams: GKS Katowice (Gruszka Piotr/Słaby Grzegorz)
#>        vs
#>        MKS Będzin (Siewiorek Emil)
#> Result: 3-2 (25-20, 18-25, 25-15, 19-25, 15-6)
#> Duration: 114 minutes

x <- ovdata_example("clickscout", as = "parsed")
summary(x)
#> Match summary:
#> Date: 2020-02-04
#> League: League Div1
#> Teams: Team XYZ ( / )
#>        vs
#>        Team ABC (Coach ABC/ )
#> Result: 3-1 (15-21, 21-19, 21-19, 23-21)
#> Duration: 112 minutes